requirements.txt
Django
psycopg2-binary
python-dotenv
PyPDF2
python-magic-bin